refactor: time data type changed to float

+++ src/main/java/com/raventool/model/response/ResponseDetails.java
@@ -5,5 +5,5 @@ import tools.jackson.databind.JsonNode;
 public record ResponseDetails(
     int status,
     JsonNode body,
-    float time
+    long time
 ) { }
